add: financial summary gql type
